Factor out ShsTiers calculation specs too
